Redesign after rigorous research

+++ make.sh
@@ -7,14 +7,16 @@ IFS=$(echo -en "\n\b") #internal field seperator
 
 dir=~/dotfiles
 oldDir=~/dotfiles/dotfiles_old
-files=( ".bash_aliases" ".bashrc" ".vimrc" ".zshrc" ".gitconfig" ".tmux.conf" )
+files=( ".aliases" ".bashrc" ".bash_profile" ".bash_prompt" ".exports" ".vimrc" ".zshrc" ".tmux.conf" ".gitconfig" )
 sublimeFiles=(
 	"Preferences.sublime-settings"
 	"Package Control.sublime-settings"
 	"Default (Linux).sublime-keymap"
+	"Default (Linux).sublime-mousemap"
 	"Anaconda.sublime-settings"
 	"Distraction Free.sublime-settings"
 	"Markdown.sublime-settings"
+	"Python-3.sublime-build"
 	)
 vsFiles=( "settings.json" )
 
@@ -29,25 +31,40 @@ cd $dir
 
 echo "Moving existing dotfiles from ~ to $oldDir"
 for file in "${files[@]}"; do
-	mv ~/$file $oldDir/
-	echo "Creating symlink to $file"
-	ln -s $dir/$file ~/$file
+	if [ -f ~/$file ]; then
+		mv ~/$file $oldDir/
+	fi
+	if [ $file == ".gitconfig" ]; then
+		echo "Copying $file file";
+		cp $dir/$file ~/$file
+	else
+		echo "Creating symlink to $file"
+		ln -s $dir/$file ~/$file
+	fi
 done
 
 echo "Moving existing dotfiles from ~/.config/sublime-text-3/Packages/User to $oldDir/sublime"
 for file in "${sublimeFiles[@]}"; do
-	mv ~/.config/sublime-text-3/Packages/User/$file $oldDir/sublime/
+	if [ -f ~/.config/sublime-text-3/Packages/User/$file ]; then
+		mv ~/.config/sublime-text-3/Packages/User/$file $oldDir/sublime/
+	fi
 	echo "Creating symlink to $file"
 	ln -s $dir/sublime/$file ~/.config/sublime-text-3/Packages/User/$file
 done
 
 echo "Moving existing dotfiles from ~/.config/Code/User to $oldDir/vscode"
 for file in "${vsFiles[@]}"; do
-	mv ~/.config/Code/User/$file $oldDir/vscode/
+	if [ -f ~/.config/Code/User/$file ]; then
+		mv ~/.config/Code/User/$file $oldDir/vscode/
+	fi
 	echo "Creating symlink to $file"
 	ln -s $dir/vscode/$file ~/.config/Code/User/$file
 done
 
 # restore $IFS
 IFS=$SAVEIFS
+
+echo "Installing Vundle plugins in Vim..."
+vim +PluginInstall +qall
+
 echo -e "\e[32m*Complete*\e[0m"
